+++ nuxt.config.js
@@ -84,6 +84,7 @@ module.exports = {
   ** Plugins to load before mounting the App
   */
   plugins: [
+    '~/plugins/path',
     '~/plugins/europeana',
     '~/plugins/vue/index',
     '~/plugins/i18n.js',
