+++ plugins/vue/index.js
@@ -2,8 +2,6 @@ import Vue from 'vue';
 
 import facets from './facets';
 import i18n from './i18n';
-import path from './path';
 
 Vue.use(facets);
 Vue.use(i18n);
-Vue.use(path);
